+++ Release Notes.md
@@ -16,6 +16,7 @@ I've also adjusted the indentation of multiline function signature end parenthes
 
 ### ✨ New features
 
+* `ActionToggle` is a toggle that performs actions when it's toggled, instead of toggling the value. 
 * `Collection+EnumeratedArray` adds `enumeratedArray()` as a shorthand for `Array(self.enumerated())`. 
 * `DotAnimation` is a re-added, previously removed view, that animates a series of dots.
 * `DotAnimationText` is a re-added, previously removed view, that renders a text followed by a dot animation.